move Bot class to its own file in core folder
