+++ bot.py
@@ -26,6 +26,8 @@ import sys
 import time
 import platform
 
+from core.bot import Bot
+
 if not sys.version_info >= (3, 9):
     print('Taigabot only runs on python 3.9+')
     sys.exit(1)
@@ -34,27 +36,6 @@ print('Taigabot <https://github.com/inexist3nce/Taigabot>')
 
 sys.path += ['plugins']  # so 'import hook' works without duplication
 os.chdir(sys.path[0] or '.')  # do stuff relative to the install directory
-
-
-class Bot:
-    def __init__(self):
-        self.start_time = time.time()
-        self.config = {}
-        self.config_mtime = 0
-        self.conns = {}
-
-        # folder used to store database and logs
-        self.persist_dir = os.path.abspath('persist')
-
-    def ensure_files(self):
-        if not os.path.exists(self.persist_dir):
-            os.mkdir(self.persist_dir)
-
-        if not os.path.exists('config'):
-            print('ERROR: no config file found!')
-            print("Please rename 'config.default' to 'config' to set up your bot!")
-            sys.exit(1)
-
 
 # print debug info
 opsys = platform.platform()
